tph/requirements-django3-375.txt
asgiref>=3.2.3
beautifulsoup4>=4.8.1
certifi>=2019.11.28
chardet>=3.0.4
coreapi>=2.3.3
coreschema>=0.0.4
deprecation>=2.0.7
Django>=3.0
django-appconf>=1.0.3
django-bootstrap4>=1.0.1
-e git://github.com/arteria/django-background-tasks.git@master#egg=django-background-tasks
-e git://github.com/mitsuhisaT/django-compressor.git@six1.13#egg=django_compressor
django-filter>=2.2.0
djangorestframework>=3.10.3
drf-yasg>=1.17.0
idna>=2.8
inflection>=0.3.1
itypes>=1.1.0
Jinja2>=2.10.3
libsass>=0.19.4
Markdown>=3.1.1
MarkupSafe>=1.1.1
packaging>=19.2
pyparsing>=2.4.5
pytz>=2019.3
rcssmin>=1.0.6
requests>=2.22.0
rjsmin>=1.1.0
ruamel.yaml>=0.16.5
ruamel.yaml.clib>=0.2.0
six>=1.13.0
soupsieve>=1.9.5
sqlparse>=0.3.0
uritemplate>=3.0.0
urllib3>=1.25.7
